pyproject.toml
[project]
name = "prompti"
version = "0.11.20"
description = "Provider-agnostic asynchronous prompt engine"
authors = [
  { name = "wolvever" }, { name = "gaoxiaoyang02" }
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "pydantic==2.11.7",
  "jinja2>=3",
  "async-lru",
  "httpx[http2]>=0.25",
  "tenacity>=8",
  "aiofiles",
  "pyyaml",
  "prometheus-client",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-exporter-otlp",
  "xxhash",
  "python-dotenv>=1.1.0",
  "semantic-version>=2.10.0",
  "pytest-asyncio",
  "elasticsearch==7.17.9",
  "opensearch-py==2.3.1",
  "litellm==1.79.1",
  "boto3==1.40.27",
  "json_repair==0.48.0",
  "openai==2.7.1",
  "google-cloud-aiplatform==1.126.1",
  "loguru>=0.7.0",
  "cryptography==43.0.3"
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio"]


[tool.uv]
# uv is used for dependency management during development

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
extend-exclude = ["tests/data"]
ignore = ["PYI001"]  # Ignore untyped import rule

[tool.ruff.lint]
select = [
    "D",   # pycodestyle docstring
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "S",   # bandit
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "B018", "D"]
"examples/*" = ["B018"]

[tool.mypy]
ignore_missing_imports = true

[tool.pyright]
exclude = ["**/__pycache__"]
reportMissingImports = "warning"
reportMissingModuleSource = "none"

[tool.pyright.strict]
